[misc] "turn off monitors on workstation lock" now waits for 5 seconds user idle (#57)
this fixes issues with screensaver activated locks and allows us to turn off the monitor again after it's been turned on by mouse movement (eg. by accident; had been requested before on DonationCoder's)

In theory, T-Clock can now react on any kind if "lock" whether it's a "lock" or "switch user" that caused the logon screen to trigger. Though Windows doesn't yet allow us to turn off monitors or detect if the system is idle when it wasn't a "lock" that caused it