Clock: fix issue with file-browse dialogs and current working directory (fixes #84)
! api.ShellExecute() will now "suggest" to use T-Clock's directory
* code cleanup which relies on the new behavior
* some other minor code adjustments

File-browse dialogs will change the working directory and might fail to restore it.
There might be other cases like third-party apps that inject stuff and do the same,
so it's bad to rely on the working directory on any Windows OS (GUI apps at least).

I'm leaving an alternative approach in place, but it shouldn't be necessary to use it (unless someone wants the unreliable working directory...)